+++ examples/train_imagenette.py
@@ -5,12 +5,13 @@ from fastai.distributed import *
 from fastai.callbacks.tracker import *
 torch.backends.cudnn.benchmark = True
 import time
+from fastai.general_optimizer import *
 
 def get_data(path, size, bs, workers):
     tfms = ([
         flip_lr(p=0.5),
-        brightness(change=(0.4,0.6)),
-        contrast(scale=(0.7,1.3))
+        #brightness(change=(0.4,0.6)),
+        #contrast(scale=(0.7,1.3))
     ], [])
     return (ImageList.from_folder(path).split_by_folder(valid='val')
             .label_from_folder().transform(tfms, size=size)
@@ -27,11 +28,19 @@ def bn_and_final(m):
     l2 = [ll[i] for i in idx]
     return split_model(splits=[l1,l2])
 
+class SGDX(GeneralOptimizer):
+    def make_step(self, p, group, group_idx):
+        p.data.add_(-group['lr'], self.state[p]['mom'])
+
+class RMSpropX(GeneralOptimizer):
+    def make_step(self, p, group, group_idx):
+        p.data.addcdiv_(-group['lr'], self.state[p]['mom_buffer'], self.state[p]['alpha_buffer'].sqrt() + 1e-7)
+
 @call_parse
 def main( gpu:Param("GPU to run on", str)=None ):
     """Distributed training of Imagenette.
     Fastest multi-gpu speed is if you run with: python -m fastai.launch"""
-    tot_epochs,lr = 5,0.3
+    bs,tot_epochs,lr = 256,5,0.0003
 
     # Pick one of these
     path,size = untar_data(URLs.IMAGENETTE_160),128
@@ -41,11 +50,18 @@ def main( gpu:Param("GPU to run on", str)=None ):
     gpu = setup_distrib(gpu)
     n_gpus = num_distrib() or 1
 
-    bs = 256//n_gpus
-
     workers = min(8, num_cpus()//n_gpus)
     data = get_data(path, size, bs, workers)
-    opt_func = partial(optim.Adam, betas=(0.9,0.99), eps=0.01)
+    #opt_func = partial(optim.Adam, betas=(0.9,0.99), eps=1e-7)
+    #opt_func = partial(optim.RMSprop, alpha=0.9)
+    #opt_func = optim.SGD
+    #opt_func = partial(SGDX, stats=[AvgStatistic('avgstat', init=0, beta=0.9),])
+    #"""
+    opt_func = partial(RMSpropX, stats=[
+        AvgStatistic('mom', 0.9),
+        AvgSquare('alpha', 0.9),
+    ])
+    #"""
     #learn = (cnn_learner(data, models.xresnet50, pretrained=False, concat_pool=False, lin_ftrs=[], split_on=bn_and_final,
     learn = (Learner(data, models.xresnet50(),
              metrics=[accuracy,top_k_accuracy], wd=1e-3, opt_func=opt_func,
@@ -54,13 +70,12 @@ def main( gpu:Param("GPU to run on", str)=None ):
         .to_fp16(dynamic=True)
         #.split(bn_and_final)
     )
-    # TODO: learn.to_parallel()
-    if gpu is None: learn.model = nn.DataParallel(learn.model)
-    else:           learn.distributed(gpu)
+    if gpu is None: learn.to_parallel()
+    else:           learn.to_distributed(gpu)
 
     # Using bs 256 on single GPU as baseline, scale the LR linearly
     bs_rat = bs/256
     lr *= bs_rat
-    learn.fit_one_cycle(tot_epochs, lr, div_factor=20, pct_start=0.7, moms=(0.9,0.9))
+    learn.fit_one_cycle(tot_epochs, lr, div_factor=20, pct_start=0.5, moms=(0.9,0.9))
     learn.save('nette')