+++ fastai/basic_data.py
@@ -8,7 +8,7 @@ __all__ = ['DataBunch', 'DeviceDataLoader', 'DatasetType', 'load_data']
 old_dl_init = torch.utils.data.DataLoader.__init__
 
 def intercept_args(self, dataset, batch_size=1, shuffle=False, sampler=None, batch_sampler=None,
-                 num_workers=0, collate_fn=default_collate, pin_memory=False, drop_last=False,
+                 num_workers=0, collate_fn=default_collate, pin_memory=True, drop_last=False,
                  timeout=0, worker_init_fn=None):
     self.init_kwargs = {'batch_size':batch_size, 'shuffle':shuffle, 'sampler':sampler, 'batch_sampler':batch_sampler,
                         'num_workers':num_workers, 'collate_fn':collate_fn, 'pin_memory':pin_memory,